Added dusk testing
